Android build for Harfbuzz NG
Adding these two files makes the new Harfbuzz library build on Android,
and creates a library suitable for integration with TextLayoutCache
(there are separate CL's for that and Skia to complete the integration).

The hb-version commit hardcodes the version and is a little hackish.
Upstream might come up with a better solution for this.

Change-Id: I87d73ae843fd8154a4b6a6f1753ee4d099339567